backend/requirements.txt
# PSWEB Python - Dependências
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fdb==2.0.2
reportlab==4.0.7
openpyxl==3.1.2
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
pywebview==4.4.1
